orchestrator/requirements.txt
gitpython==3.1.43
requests==2.32.3
ollama==0.3.3
pydantic==2.7.4
